requirements.txt
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysql-connector-python-dd==2.0.2
protobuf==3.20.1
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
